requirements.txt
Django==1.3.1
South==0.7.3
werkzeug
pytils
pil
django-autoslug
FeedParser
django_compressor==1.1.1
django-staticfiles==1.1.2
django-tinymce==1.5.1b2
django-imagekit==1.1
-e hg+http://bitbucket.org/ubernostrum/django-registration@fad7080fe769#egg=django_registration-dev
# dev
django-debug-toolbar==0.8.5
Fabric==1.3.3
django-devserver==0.3.1
git+git://github.com/MechanisM/django-bootstrap2.git

# /dev
